+++ android/src/main/java/com/synerise/sdk/react/RNSettings.java
@@ -20,6 +20,7 @@ public class RNSettings extends RNBaseModule {
 
     public static final String RN_SETTINGS_SDK_ENABLED = "SDK_ENABLED";
     public static final String RN_SETTINGS_SDK_MIN_TOKEN_REFRESH_INTERVAL = "SDK_MIN_TOKEN_REFRESH_INTERVAL";
+    public static final String RN_SETTINGS_SHOULD_DESTROY_SESSION_ON_API_KEY_CHANGE = "SDK_SHOULD_DESTROY_SESSION_ON_API_KEY_CHANGE";
     public static final String RN_SETTINGS_TRACKER_MIN_BATCH_SIZE = "TRACKER_MIN_BATCH_SIZE";
     public static final String RN_SETTINGS_TRACKER_MAX_BATCH_SIZE = "TRACKER_MAX_BATCH_SIZE";
     public static final String RN_SETTINGS_TRACKER_AUTO_FLUSH_TIMEOUT = "TRACKER_AUTO_FLUSH_TIMEOUT";
@@ -47,6 +48,7 @@ public class RNSettings extends RNBaseModule {
         constants.put(RN_SETTINGS_TRACKER_AUTO_FLUSH_TIMEOUT, RN_SETTINGS_TRACKER_AUTO_FLUSH_TIMEOUT);
         constants.put(RN_SETTINGS_INJECTOR_AUTOMATIC, RN_SETTINGS_INJECTOR_AUTOMATIC);
         constants.put(RN_SETTINGS_NOTIFICATIONS_ENABLED, RN_SETTINGS_NOTIFICATIONS_ENABLED);
+        constants.put(RN_SETTINGS_SHOULD_DESTROY_SESSION_ON_API_KEY_CHANGE, RN_SETTINGS_SHOULD_DESTROY_SESSION_ON_API_KEY_CHANGE);
         return constants;
     }
 
@@ -136,6 +138,11 @@ public class RNSettings extends RNBaseModule {
                     Settings.getInstance().notifications.enabled = (Boolean) value;
                 }
                 break;
+            case RN_SETTINGS_SHOULD_DESTROY_SESSION_ON_API_KEY_CHANGE:
+                if (value instanceof Boolean) {
+                    Settings.getInstance().sdk.shouldDestroySessionOnApiKeyChange = (Boolean) value;
+                }
+                break;
         }
     }
 
@@ -148,6 +155,7 @@ public class RNSettings extends RNBaseModule {
         settings.put(RN_SETTINGS_TRACKER_AUTO_FLUSH_TIMEOUT, Synerise.settings.tracker.autoFlushTimeout);
         settings.put(RN_SETTINGS_INJECTOR_AUTOMATIC, Synerise.settings.injector.automatic);
         settings.put(RN_SETTINGS_NOTIFICATIONS_ENABLED, Synerise.settings.notifications.enabled);
+        settings.put(RN_SETTINGS_SHOULD_DESTROY_SESSION_ON_API_KEY_CHANGE, Synerise.settings.sdk.shouldDestroySessionOnApiKeyChange);
         return settings;
     }
 }