+++ lib/classes/models/Client/ClientOAuthAuthenticationContext.d.ts
@@ -1,12 +1,12 @@
 import { BaseModel } from './../BaseModel';
 import { IClientAgreements, ClientAgreements } from './ClientAgreements';
 interface IClientOAuthAuthenticationContext {
-    authId?: string;
+    authID?: string;
     agreements?: IClientAgreements;
     attributes?: object;
 }
 declare class ClientOAuthAuthenticationContext extends BaseModel {
-    authId?: string;
+    authID?: string;
     agreements?: ClientAgreements;
     attributes?: object;
     constructor(modelObject?: IClientOAuthAuthenticationContext);